requirements.txt
absl-py==1.4.0
albumentations==1.3.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==2.10
imageio==2.27.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.3
joblib==1.2.0
keras==2.12.0
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.5
ml-dtypes==0.1.0
networkx==3.1
numpy>=1.19,<1.24
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.4.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy>=1.8,<1.9.2
six==1.16.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow==2.12.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.4.12
typing_extensions==4.5.0
urllib3==1.26.6
Werkzeug==3.0.1
wrapt==1.14.1
